[MS] Renamed notificationCenter to notificationManager
[MS] Removed await in front of showToast

[MS] Added missing awaits in front of showModal

[MS] Changed naming in testNotificationManager
